+++ models/User/Teacher.cs
@@ -9,9 +9,10 @@ namespace PSI_DA_PL_B.models.User
     internal class Teacher : Client
     {
         public string Email {  get; set; }
+
         public Teacher() : base() { }
 
-        public Teacher(string name, int nif, double balance, string email) : base(name, nif)
+        public Teacher(string name, int nif, decimal balance, string email) : base(name, nif)
         {
             this.Email = email;
         }
